Moved extended flat app to its own module
